backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.125.0",
    "psycopg[binary]>=3.3.2",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
    "asyncpg>=0.29.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
